+++ Inc/saiportadapter.hpp
@@ -92,10 +92,33 @@ class SaiPortAdapter : public AudioPortAdapterStrategy {
     virtual unsigned int GetNumberOfChannelsTx();
 
     /**
-     * @brief Return the size of the one sample in I2S frame.
-     * @return 1..32. The unit is [bits]
+     * @brief Return the bit count to shift to make the DMA data to right align in TX I2S frame.
+     * @return Always 0.
+     * @details
+     * This is needed because of the mismatch in the DMA buffer data formant and I2S format.
+     *
+     * Let's assume the 24bit data I2S format. Some peripheral place the data as
+     * right aligned in 32bit DMA data ( as integer ), some peripheral places the data
+     * as left aligned in 32bit DMA data ( as fixed point ).
+     *
+     * This kind of the mismatch will be aligned by audio frame work. This member function returns
+     * how many bits have to be shifted to right in TX.
+     *
+     * If peripheral requires left align format, this function shuld return 0.
+     *
+     * The STM32 SAI DMA format is right aligned. 
      */
-    virtual unsigned int GetSampleDataSizeTx();
+    virtual unsigned int GetSampleShiftSizeTx();
+
+    /**
+     * @brief Return the size of the one sample on memory for Tx channel
+     * @return 2 or 4. The unit is [Byte]
+     * @details
+     * This function returns the size of the word which should be
+     * allocated on the memory.
+     */
+    virtual unsigned int GetSampleWordSizeTx();
+
     /**
      * @brief Return how many channels are in the transfer.
      * @return 1 for Mono, 2 for stereo, 3... for multi-channel.
@@ -103,10 +126,32 @@ class SaiPortAdapter : public AudioPortAdapterStrategy {
     virtual unsigned int GetNumberOfChannelsRx();
 
     /**
-     * @brief Return the size of the one sample in I2S flame.
-     * @return 1..32. The unit is [bits]
+     * @brief Return the bit count to shift to make the DMA data to right align in RX I2S frame.
+     * @return 0 The unit is [bit]
+     * @details
+     * This is needed because of the mismatch in the DMA buffer data formant and I2S format.
+     *
+     * Let's assume the 24bit data I2S format. Some peripheral place the data as
+     * right aligned in 32bit DMA data ( as integer ), some peripheral places the data
+     * as left aligned in 32bit DMA data ( as fixed point ).
+     *
+     * This kind of the mismatch will be aligned by audio frame work. This member function returns
+     * how many bits have to be shifted to left in RX.
+     *
+     * If peripheral requires left align format, this function shuld return 0.
+     *
+     * The STM32 I2S DMA format is left aligned. So, always return 0.
      */
-    virtual unsigned int GetSampleDataSizeRx();
+    virtual unsigned int GetSampleShiftSizeRx();
+
+    /**
+     * @brief Return the size of the one sample on memory for Rx channel
+     * @return 2 or 4. The unit is [Byte]
+     * @details
+     * This function returns the size of the word which should be
+     * allocated on the memory.
+     */
+    virtual unsigned int GetSampleWordSizeRx();
     /**
      * @brief Handling error report of device.
      * @param ptr Pointer for generic use. Usually, points a struct of a device control