+++ Src/saiportadapter.cpp
@@ -147,7 +147,7 @@ unsigned int SaiPortAdapter::GetNumberOfChannelsRx()
     return return_val;
 }
 
-unsigned int SaiPortAdapter::GetSampleDataSizeRx()
+unsigned int SaiPortAdapter::GetSampleShiftSizeRx()
 {
     SAIAUDIO_SYSLOG("Enter.")
     MURASAKI_ASSERT(rx_peripheral_ != nullptr)
@@ -156,22 +156,58 @@ unsigned int SaiPortAdapter::GetSampleDataSizeRx()
 
     switch (rx_peripheral_->Init.DataSize) {
         case SAI_DATASIZE_8:
-            return_val = 8;
+            return_val = 8;  // The DMA word size is 2 byte. So, must shift 8 bit.
             break;
         case SAI_DATASIZE_10:
-            return_val = 10;
+            return_val = 6;  // The DMA word size is 2 byte. So, must shift 6 bit.
             break;
         case SAI_DATASIZE_16:
-            return_val = 16;
+            return_val = 0;
             break;
         case SAI_DATASIZE_20:
-            return_val = 20;
+            return_val = 12;  // The DMA word size is 4 byte. So, must shift 12bit.
             break;
         case SAI_DATASIZE_24:
-            return_val = 24;
+            return_val = 8;  // The DMA word size is 4 byte. So, must shift 8bit.
             break;
         case SAI_DATASIZE_32:
-            return_val = 32;
+            return_val = 0;
+            break;
+        default:
+            MURASAKI_SYSLOG(this, kfaSai, kseError, "Unexpected data size")
+            MURASAKI_ASSERT(false)
+            // force assertion.
+    }
+
+    SAIAUDIO_SYSLOG("Exit with %d.", return_val)
+    return return_val;
+}
+
+unsigned int SaiPortAdapter::GetSampleWordSizeRx()
+{
+    SAIAUDIO_SYSLOG("Enter.")
+    MURASAKI_ASSERT(rx_peripheral_ != nullptr)
+
+    unsigned int return_val;
+
+    switch (rx_peripheral_->Init.DataSize) {
+        case SAI_DATASIZE_8:
+            return_val = 2;  // The DMA word size is 2 byte.
+            break;
+        case SAI_DATASIZE_10:
+            return_val = 2;  // The DMA word size is 2 byte.
+            break;
+        case SAI_DATASIZE_16:
+            return_val = 2;
+            break;
+        case SAI_DATASIZE_20:
+            return_val = 4;  // The DMA word size is 4 byte.
+            break;
+        case SAI_DATASIZE_24:
+            return_val = 4;  // The DMA word size is 4 byte.
+            break;
+        case SAI_DATASIZE_32:
+            return_val = 4;
             break;
         default:
             MURASAKI_SYSLOG(this, kfaSai, kseError, "Unexpected data size")
@@ -194,7 +230,7 @@ unsigned int SaiPortAdapter::GetNumberOfChannelsTx()
     return return_val;
 }
 
-unsigned int SaiPortAdapter::GetSampleDataSizeTx()
+unsigned int SaiPortAdapter::GetSampleShiftSizeTx()
 {
     SAIAUDIO_SYSLOG("Enter.")
     MURASAKI_ASSERT(tx_peripheral_ != nullptr)
@@ -203,22 +239,58 @@ unsigned int SaiPortAdapter::GetSampleDataSizeTx()
 
     switch (rx_peripheral_->Init.DataSize) {
         case SAI_DATASIZE_8:
-            return_val = 8;
+            return_val = 8;  // The DMA word size is 2 byte. So, must shift 8 bit.
             break;
         case SAI_DATASIZE_10:
-            return_val = 10;
+            return_val = 6;  // The DMA word size is 2 byte. So, must shift 6 bit.
             break;
         case SAI_DATASIZE_16:
-            return_val = 16;
+            return_val = 0;
             break;
         case SAI_DATASIZE_20:
-            return_val = 20;
+            return_val = 12;  // The DMA word size is 4 byte. So, must shift 12bit.
             break;
         case SAI_DATASIZE_24:
-            return_val = 24;
+            return_val = 8;  // The DMA word size is 4 byte. So, must shift 8bit.
             break;
         case SAI_DATASIZE_32:
-            return_val = 32;
+            return_val = 0;
+            break;
+        default:
+            MURASAKI_SYSLOG(this, kfaSai, kseError, "Unexpected data size")
+            MURASAKI_ASSERT(false)
+            // force assertion.
+    }
+
+    SAIAUDIO_SYSLOG("Exit with %d.", return_val)
+    return return_val;
+}
+
+unsigned int SaiPortAdapter::GetSampleWordSizeTx()
+{
+    SAIAUDIO_SYSLOG("Enter.")
+    MURASAKI_ASSERT(tx_peripheral_ != nullptr)
+
+    unsigned int return_val;
+
+    switch (rx_peripheral_->Init.DataSize) {
+        case SAI_DATASIZE_8:
+            return_val = 2;  // The DMA word size is 2 byte.
+            break;
+        case SAI_DATASIZE_10:
+            return_val = 2;  // The DMA word size is 2 byte.
+            break;
+        case SAI_DATASIZE_16:
+            return_val = 2;
+            break;
+        case SAI_DATASIZE_20:
+            return_val = 4;  // The DMA word size is 4 byte.
+            break;
+        case SAI_DATASIZE_24:
+            return_val = 4;  // The DMA word size is 4 byte.
+            break;
+        case SAI_DATASIZE_32:
+            return_val = 4;
             break;
         default:
             MURASAKI_SYSLOG(this, kfaSai, kseError, "Unexpected data size")